@@ -1,20 +1,14 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
-    <AspNetCoreHostingModel>InProcess</AspNetCoreHostingModel>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
   </PropertyGroup>
-
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.App" />
-    <PackageReference Include="Microsoft.AspNetCore.Razor.Design" Version="2.2.0" PrivateAssets="All" />
-    <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="2.2.4" />
+  
+   <ItemGroup>
+         <PackageReference Include="Microsoft.Extensions.Logging.Debug" Version="3.1.0" />
+         <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="3.1.0" />
+         <PackageReference Include="Syncfusion.EJ2.AspNet.Core" Version="17.4.0.44" />
   </ItemGroup>
 
-
-  <ItemGroup>
-    <Folder Include="Views\Products\" />
-  </ItemGroup>
 
 </Project>
