@@ -1,4 +1,6 @@
-﻿using InGame.Api.Models;
+﻿using Ext.Net;
+using InGame.Api.Data;
+using InGame.Api.Models;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -8,22 +10,65 @@ namespace InGame.Api.DataService
 {
     public interface ICategoryDataService
     {
-        List<Category> GetProducts();
-        List<Product> GetProductsByCategory(int categoryId);
+        List<Category> GetCategories();
+        Product GetCategory(int categoryId);
+        bool CategoryExists(int categoryId);
 
-        Product GetProduct(int productId);
-        bool ProductExists(int productId);
-
-        bool IsDuplicateProductName(int countryId, string countryName);
-        bool CreateProduct(Product product);
-        bool UpdateProduct(Product product);
-        bool DeleteProduct(Product product);
+        bool IsDuplicateCategoryName(int countryId, string countryName);
+        bool CreateCategory(Category category);
+        bool UpdateCategory(Category category);
+        bool DeleteCategory(Category category);
         bool Save();
     }
-    
 
-
-    public class CategoryDataService
+    public class CategoryDataService : ICategoryDataService
     {
+        private InGameDataContext _context;
+        public CategoryDataService(InGameDataContext context)
+        {
+            _context = context;
+        }
+
+        public List<Category> GetCategories()
+        {
+            var result = _context.Categories.ToList();
+            return result;
+        }
+
+        public Product GetCategory(int categoryId)
+        {
+            throw new NotImplementedException();
+        }
+
+        public bool CategoryExists(int categoryId)
+        {
+            throw new NotImplementedException();
+        }
+
+        public bool CreateCategory(Category category)
+        {
+            throw new NotImplementedException();
+        }
+
+        public bool DeleteCategory(Category category)
+        {
+            throw new NotImplementedException();
+        }
+ 
+        public bool IsDuplicateCategoryName(int countryId, string countryName)
+        {
+            throw new NotImplementedException();
+        }
+
+        public bool Save()
+        {
+            throw new NotImplementedException();
+        }
+
+        public bool UpdateCategory(Category category)
+        {
+            throw new NotImplementedException();
+        }
+
     }
 }
