@@ -1,35 +1,25 @@
 ﻿<!DOCTYPE html>
-
 <html lang="en">
-
-<!-- begin::Head -->
 <head>
-    <base href="../../">
     <meta charset="utf-8" />
-    <title>Metronic | Empty Page</title>
-    <meta name="description" content="Page with empty content">
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>@ViewData["Title"] - InGame.WebUi</title>
+    <link rel="stylesheet" href="~/lib/bootstrap/dist/css/bootstrap.min.css" />
+    <link rel="stylesheet" href="~/css/site.css" />
+    <link rel="stylesheet" href="~/css/ej2/material.css" />
 
     <!--begin::Fonts -->
     <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Poppins:300,400,500,600,700|Roboto:300,400,500,600,700">
-
     <!--end::Fonts -->
-    <!--begin::Page Vendors Styles(used by this page) -->
-    <link href="~/assets/plugins/custom/fullcalendar/fullcalendar.bundle.css" rel="stylesheet" type="text/css" />
-
-    <!--end::Page Vendors Styles -->
     <!--begin::Global Theme Styles(used by all pages) -->
     <link href="~/assets/plugins/global/plugins.bundle.css" rel="stylesheet" type="text/css" />
     <link href="~/assets/css/style.bundle.css" rel="stylesheet" type="text/css" />
-
     <!--end::Global Theme Styles -->
     <!--begin::Layout Skins(used by all pages) -->
     <!--end::Layout Skins -->
     <link rel="shortcut icon" href="~/assets/media/logos/favicon.ico" />
+    <script src="~/js/ej2/ej2.min.js"></script>
 </head>
-
-<!-- end::Head -->
-<!-- begin::Body -->
 <body class="kt-quick-panel--right kt-demo-panel--right kt-offcanvas-panel--right kt-header--fixed kt-header-mobile--fixed kt-subheader--enabled kt-subheader--solid kt-aside--enabled kt-aside--fixed kt-aside--minimize kt-page--loading">
 
     <!-- begin:: Page -->
@@ -142,10 +132,9 @@
 
                     <!-- end:: Header Topbar -->
                 </div>
-
                 <!-- end:: Header -->
                 <div class="kt-content  kt-grid__item kt-grid__item--fluid kt-grid kt-grid--hor" id="kt_content">
-                    
+
                     <!-- begin:: Content -->
                     <div class="kt-container  kt-container--fluid  kt-grid__item kt-grid__item--fluid">
                         @RenderBody()
@@ -178,6 +167,7 @@
     <div id="kt_scrolltop" class="kt-scrolltop">
         <i class="fa fa-arrow-up"></i>
     </div>
+
     <!-- end::Scrolltop -->
     <!-- begin::Global Config(global config for global JS sciprts) -->
     <script>
@@ -201,6 +191,12 @@
         };
     </script>
 
+    <script src="~/lib/jquery/dist/jquery.min.js"></script>
+    <script src="~/lib/bootstrap/dist/js/bootstrap.bundle.min.js"></script>
+    <script src="~/js/site.js" asp-append-version="true"></script>
+    <ejs-scripts></ejs-scripts>
+
+
     <!-- end::Global Config -->
     <!--begin::Global Theme Bundle(used by all pages) -->
     <script src="~/assets/plugins/global/plugins.bundle.js" type="text/javascript"></script>
@@ -215,9 +211,8 @@
     <!--end::Page Vendors -->
     <!--begin::Page Scripts(used by this page) -->
     <script src="~/assets/js/pages/dashboard.js" type="text/javascript"></script>
-
     <!--end::Page Scripts -->
-</body>
 
-<!-- end::Body -->
+    @RenderSection("Scripts", required: false)
+</body>
 </html>
